Make custom font sizes appear fluid in the block editor when fluid typography is enabled
